requirements.txt
fastapi
uvicorn
python-dotenv
pymupdf
langchain
langchain-openai
openai


